+++ source/scripts/cmake/fn_python3_external.cmake
@@ -1,7 +1,14 @@
 function(python3_external)
     set(options)
     set(oneValueArgs PROJECT_NAME BUILD_VARIANT)
-    set(multiValueArgs)
+    set(multiValueArgs
+        INCLUDE_DIRS
+        LINK_LIBS
+        LINK_DIRS
+        COMPILE_DEFINITIONS
+        COMPILE_OPTIONS
+        LINK_OPTIONS
+    )
 
     cmake_parse_arguments(
         PY3EXT
@@ -11,7 +18,17 @@ function(python3_external)
         ${ARGN}
     )
 
+    message(STATUS "PROJECT_NAME: ${PY3EXT_PROJECT_NAME}")
+    message(STATUS "BUILD_VARIANT: ${PY3EXT_BUILD_VARIANT}")
+    # message(STATUS "PY3EXT_INCLUDE_DIRS: ${PY3EXT_INCLUDE_DIRS}")
+    # message(STATUS "PY3EXT_COMPILE_DEFINITIONS: ${PY3EXT_COMPILE_DEFINITIONS}")
+    # message(STATUS "PY3EXT_COMPILE_OPTIONS: ${PY3EXT_COMPILE_OPTIONS}")
+    # message(STATUS "PY3EXT_LINK_LIBS: ${PY3EXT_LINK_LIBS}")
+    # message(STATUS "PY3EXT_LINK_DIRS: ${PY3EXT_LINK_DIRS}")
+    # message(STATUS "PY3EXT_LINK_OPTIONS: ${PY3EXT_LINK_OPTIONS}")
+
     set(variants local shared-ext static-ext framework-ext framework-pkg)
+    set(self_contained shared-ext static-ext framework-ext)
 
     if (NOT ${PY3EXT_BUILD_VARIANT} IN_LIST variants)
         message(FATAL_ERROR 
@@ -28,9 +45,6 @@ function(python3_external)
     elseif(PY3EXT_BUILD_VARIANT STREQUAL "framework-pkg")
         set(BUILD_FRAMEWORK_PKG 1)
     endif()
-
-    message(STATUS "PROJECT_NAME: ${PY3EXT_PROJECT_NAME}")
-    message(STATUS "BUILD_VARIANT: ${PY3EXT_BUILD_VARIANT}")
 
     set(DEPS_DIR "${CMAKE_SOURCE_DIR}/build/install")
     set(EXTERNAL_DIR "${CMAKE_LIBRARY_OUTPUT_DIRECTORY}/${PY3EXT_PROJECT_NAME}.mxo")
@@ -84,6 +98,7 @@ function(python3_external)
         ${PY3EXT_PROJECT_NAME}
         PRIVATE
         ${Python3_INCLUDE_DIRS}
+        ${PY3EXT_INCLUDE_DIRS}
         $<$<BOOL:${BUILD_STATIC_EXT}>:${DEPS_DIR}/bzip2/include>
         $<$<BOOL:${BUILD_STATIC_EXT}>:${DEPS_DIR}/openssl/include>
         $<$<BOOL:${BUILD_STATIC_EXT}>:${DEPS_DIR}/xz/include>
@@ -92,13 +107,16 @@ function(python3_external)
     target_compile_definitions(
         ${PY3EXT_PROJECT_NAME}
         PRIVATE
+        ${PY3EXT_COMPILE_DEFINITIONS}
         -DNDEBUG
         $<$<BOOL:${BUILD_STATIC_EXT}>:-DBUILD_STATIC> # help static find pyhome
+        # $<IN_LIST:${PY3EXT_BUILD_VARIANT},${self_contained}:-DSELFCONTAINED_EXTERNAL> # special case
     )
 
     target_compile_options(
         ${PY3EXT_PROJECT_NAME}
         PRIVATE
+        ${PY3EXT_COMPILE_OPTIONS}
         $<$<PLATFORM_ID:Darwin>:-Wno-unused-result>
         $<$<PLATFORM_ID:Darwin>:-Wsign-compare>
         $<$<PLATFORM_ID:Darwin>:-Wunreachable-code>
@@ -112,6 +130,7 @@ function(python3_external)
         ${PY3EXT_PROJECT_NAME} 
         PRIVATE
         ${Python3_LIBRARY_DIRS}
+        ${PY3EXT_LINK_DIRS}
         $<$<BOOL:${BUILD_STATIC_EXT}>:${DEPS_DIR}/bzip2/lib>
         $<$<BOOL:${BUILD_STATIC_EXT}>:${DEPS_DIR}/openssl/lib>
         $<$<BOOL:${BUILD_STATIC_EXT}>:${DEPS_DIR}/xz/lib>
@@ -127,10 +146,17 @@ function(python3_external)
         "$<$<PLATFORM_ID:Darwin>:-lsqlite3>"
     )
 
+    target_link_options(
+        ${PY3EXT_PROJECT_NAME}
+        PRIVATE
+        ${PY3EXT_LINK_OPTIONS}
+    )
+
     target_link_libraries(
         ${PY3EXT_PROJECT_NAME} 
         PRIVATE
         "${Python3_LIBRARIES}"
+        "${PY3EXT_LINK_LIBS}"
         "$<$<PLATFORM_ID:Darwin>:-ldl>"
         "$<$<PLATFORM_ID:Darwin>:-framework CoreFoundation>"
         "$<$<BOOL:${BUILD_STATIC_EXT}>:${STATIC_LINK_DEPS}>"