+++ source/projects/ztp/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 [0.1.0]
 
-- Renamed this project from `zthread` to `ztp`
+- Changed name of this project from `zthread` to `ztp`
 
 - Successful proof-of-concept project created to demonstrate accessing a spawned python interpreter via the use of threads and zeromq channels in order to prevent blocking.